refactor: 도메인 패키지 역할별 구조화
