+++ src/test/kotlin/blackjack/domain/model/participant/PlayerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.model
+package blackjack.domain.model.participant
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
