+++ src/test/kotlin/blackjack/domain/model/progress/RuleTest.kt
@@ -1,5 +1,6 @@
-package blackjack.domain.model
+package blackjack.domain.model.progress
 
+import blackjack.domain.model.card.Card
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
